requirements.txt
# AI
google-generativeai>=0.8.0
Pillow>=10.0.0

# Desktop App (cross-platform)
pystray>=0.19.0
pynput>=1.7.0
pyperclip>=1.9.0
pyautogui>=0.9.54

# macOS only — Accessibility API (auto-skipped on Windows)
# Install with: pip install pyobjc-core pyobjc-framework-ApplicationServices pyobjc-framework-Quartz
# pyobjc-core>=10.0
# pyobjc-framework-ApplicationServices>=10.0
# pyobjc-framework-Quartz>=10.0

# Windows only — UI Automation (auto-skipped on macOS)
# Install with: pip install uiautomation psutil mss
# uiautomation>=2.0.18
# psutil>=5.9.0
# mss>=9.0.0

# Gmail (optional — for main.py bot mode)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Telegram Bot (optional — for main.py bot mode)
python-telegram-bot>=21.0

# Database
aiosqlite>=0.20.0

# Scheduler
apscheduler>=3.10.0

# Config
python-dotenv>=1.0.0

# Async
httpx>=0.27.0
